Move `ScoreMove` and History/Killer shenanigans to a separated file, to leave NegaMax cleaner
