+++ src/Lynx/Search/Helpers.cs
@@ -8,124 +8,8 @@ namespace Lynx;
 
 public sealed partial class Engine
 {
-    /// <summary>
-    /// Returns the score evaluation of a move taking into account <paramref name="bestMoveTTCandidate"/>, <see cref="EvaluationConstants.MostValueableVictimLeastValuableAttacker"/>, <see cref="_killerMoves"/> and <see cref="_quietHistory"/>
-    /// </summary>
-    /// <param name="move"></param>
-    /// <param name="ply"></param>
-    /// <param name="isNotQSearch"></param>
-    /// <param name="bestMoveTTCandidate"></param>
-    /// <returns></returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal int ScoreMove(Move move, int ply, bool isNotQSearch, ShortMove bestMoveTTCandidate = default)
-    {
-        if ((ShortMove)move == bestMoveTTCandidate)
-        {
-            return EvaluationConstants.TTMoveScoreValue;
-        }
-
-        var promotedPiece = move.PromotedPiece();
-        var isPromotion = promotedPiece != default;
-        var isCapture = move.IsCapture();
-
-        // Queen promotion
-        if ((promotedPiece + 2) % 6 == 0)
-        {
-            var baseScore = SEE.HasPositiveScore(Game.CurrentPosition, move)
-                ? EvaluationConstants.GoodCaptureMoveBaseScoreValue
-                : EvaluationConstants.BadCaptureMoveBaseScoreValue;
-
-            var captureBonus = isCapture ? 1 : 0;
-
-            return baseScore + EvaluationConstants.PromotionMoveScoreValue + captureBonus;
-        }
-
-        if (isCapture)
-        {
-            var baseCaptureScore = (isPromotion || move.IsEnPassant() || SEE.IsGoodCapture(Game.CurrentPosition, move))
-                ? EvaluationConstants.GoodCaptureMoveBaseScoreValue
-                : EvaluationConstants.BadCaptureMoveBaseScoreValue;
-
-            var piece = move.Piece();
-            var capturedPiece = move.CapturedPiece();
-
-            Debug.Assert(capturedPiece != (int)Piece.K && capturedPiece != (int)Piece.k, $"{move.UCIString()} capturing king is generated in position {Game.CurrentPosition.FEN()}");
-
-            return baseCaptureScore
-                + EvaluationConstants.MostValueableVictimLeastValuableAttacker[piece][capturedPiece]
-                //+ EvaluationConstants.MVV_PieceValues[capturedPiece]
-                + _captureHistory[CaptureHistoryIndex(piece, move.TargetSquare(), capturedPiece)];
-        }
-
-        if (isPromotion)
-        {
-            return EvaluationConstants.PromotionMoveScoreValue;
-        }
-
-        if (isNotQSearch)
-        {
-            var thisPlyKillerMoves = _killerMoves[ply];
-
-            // 1st killer move
-            if (thisPlyKillerMoves[0] == move)
-            {
-                return EvaluationConstants.FirstKillerMoveValue;
-            }
-
-            // 2nd killer move
-            if (thisPlyKillerMoves[1] == move)
-            {
-                return EvaluationConstants.SecondKillerMoveValue;
-            }
-
-            // 3rd killer move
-            if (thisPlyKillerMoves[2] == move)
-            {
-                return EvaluationConstants.ThirdKillerMoveValue;
-            }
-
-            if (ply >= 1)
-            {
-                var previousMove = Game.PopFromMoveStack(ply - 1);
-                Debug.Assert(previousMove != 0);
-                var previousMovePiece = previousMove.Piece();
-                var previousMoveTargetSquare = previousMove.TargetSquare();
-
-                // Countermove
-                if (_counterMoves[CounterMoveIndex(previousMovePiece, previousMoveTargetSquare)] == move)
-                {
-                    return EvaluationConstants.CounterMoveValue;
-                }
-
-                // Counter move history
-                return EvaluationConstants.BaseMoveScore
-                    + _quietHistory[move.Piece()][move.TargetSquare()]
-                    + _continuationHistory[ContinuationHistoryIndex(move.Piece(), move.TargetSquare(), previousMovePiece, previousMoveTargetSquare, 0)];
-            }
-
-            // History move or 0 if not found
-            return EvaluationConstants.BaseMoveScore
-                + _quietHistory[move.Piece()][move.TargetSquare()];
-        }
-
-        return EvaluationConstants.BaseMoveScore;
-    }
-
-    /// <summary>
-    /// Soft caps history score
-    /// Formula taken from EP discord, https://discord.com/channels/1132289356011405342/1132289356447625298/1141102105847922839
-    /// </summary>
-    /// <param name="score"></param>
-    /// <param name="rawHistoryBonus"></param>
-    /// <returns></returns>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static int ScoreHistoryMove(int score, int rawHistoryBonus)
-    {
-        return score + rawHistoryBonus - (score * Math.Abs(rawHistoryBonus) / Configuration.EngineSettings.History_MaxMoveValue);
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public void PrefetchTTEntry()
+    private void PrefetchTTEntry()
     {
         if (Sse.IsSupported)
         {
